requirements.txt
altair==5.1.1
faiss-cpu==1.7.4
hdbscan==0.8.1
numba==0.58.0
numpy==1.25.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
streamlit==1.27.0
tokenizers==0.13.3
torch==2.0.1
transformers==4.33.2
umap-learn==0.5.4